Update complex args tests
